Added a whole bunch of components and started got rid of the companyinfo box
This box really served no purpose, so instead of holding on to it, Trine and I
decided to get rid of it entirely. This has prompted some changes in how/where
boxes are placed and sized. This commit represents some of the first changes
required due to no longer having the company info box.