Found a focus related bug. Fixed now.
Lesson learned. In order for the focus system to work, we need
to register every element that should have focus, and for those
who we don't want it, but the element have an implicit structure
to get it. like input elements, then we need to explicit tell it
not to get focus with tabindex="-1"