Removed stale TODO.
A sequencer should only be added, if the dialplans were to be stored as
files identified by a unique id.
As of now, they are uniquely identified by their extention (a string).
The only potential problem this could cause (and the problem remedied
by using integer indexes) would be that a trashing of a previously
trahed dialplan (with the same extension) would either overwrite the
previously trashed dialplan or merely fail, because the file in the
trash folder already existed.

This problem is handled by appending a timestamp to the filename of
trashed files instead, rendering the issue of the TODO invalid.